fix(apps): one row per app, and routing that survives new streams (#44)
Apps are now one row carrying all of their streams (a browser opens one
per playing tab), routing a row moves every stream of the app, and the
header counts apps rather than streams.

Auto-route enforcement moves out of the UI poll into a backend ticker
(200ms native, 2s pactl with a UI-poll grace), so assignments hold while
Sink sits in the tray and a browser rebuilding its stream is back on its
channel in ~200ms instead of blasting the default output. The
already-handled ledger keys on object.serial instead of recycled node
ids, so a stream reopened between polls is routed again.

Maintainer changes on top of the original: stream refreshes are
serialized behind a gate (concurrent passes could race their seen.json
saves and un-ledger fresh streams), route_app_to_channel records intent
before moving (a concurrent tick could undo the move), the ticker backs
off tenfold after three straight failures, split apps are labeled
"N of M streams", and comments were trimmed to the repo's why-only rule.

Thanks @cody-young for the diagnosis and the fix - all three root causes
were real and verified end to end.

+++ src-tauri/src/mixer/state.rs
@@ -42,10 +42,9 @@ pub struct MixerState {
     pub buses: crate::persistence::buses::Buses,
     /// App preferences (device naming etc.), persisted to disk.
     pub prefs: crate::persistence::prefs::Prefs,
-    /// Stream indices already considered for auto-routing this session.
-    /// Each stream is enforced once, on first sight, so a user moving a
-    /// stream elsewhere (here or in pavucontrol) isn't fought every poll.
-    pub auto_routed: HashSet<u32>,
+    /// Streams already auto-routed once, by `object.serial` (node ids
+    /// recycle); manual re-routing isn't fought.
+    pub auto_routed: HashSet<u64>,
 }
 
 impl MixerState {
